Adding 2 candidate data sets
Hey Team!

I found these 2 data sets that could be good candates for our project.

I found the Obesity very interesting but had a hard time finding the correlations, as many of the features are categorical - I'm not that experienced, but maybe you guys are better than me on that.

My favorite is the Combined Cycle Power Plant. Seems like a very straight forward example to use, meets our criteria, it has good correlations in it, I think can be solved with regression, and I also find it interesting - to predict the electric output of a power plant based on several factors. Let's discuss tomorrow in our meeting.